moved GetChainIdFromConnectionId
